excel-agent-backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pandas==2.3.2
numpy==2.3.2
plotly==6.3.0
google-generativeai==0.8.5
pydantic==2.11.7
langgraph==0.6.7
langchain-core==0.3.75
python-dotenv==1.1.1
azure-ai-inference
azure-core
langchain-nvidia-ai-endpoints
